requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --allow-unsafe --output-file=requirements-dev.txt requirements-dev.in
#
astor==0.8.1
    # via flake8-simplify
astpretty==2.1.0
    # via flake8-expression-complexity
astroid==2.8.5
    # via pylint
asttokens==2.0.5
    # via flake8-aaa
attrs==21.2.0
    # via
    #   flake8-bugbear
    #   flake8-eradicate
    #   pytest
autoflake==1.4
    # via autoimport
autoimport==0.7.5
    # via -r requirements-dev.in
backports.entry-points-selectable==1.1.1
    # via virtualenv
bandit==1.7.1
    # via -r requirements-dev.in
black==21.10b0
    # via -r requirements-dev.in
certifi==2021.10.8
    # via
    #   -c docs/requirements.txt
    #   -c requirements.txt
    #   requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.7
    # via
    #   -c docs/requirements.txt
    #   -c requirements.txt
    #   requests
click==8.0.3
    # via
    #   -c docs/requirements.txt
    #   -c requirements.txt
    #   autoimport
    #   black
    #   pip-tools
    #   safety
    #   yamlfix
colorama==0.4.4
    # via flakehell
coverage[toml]==6.1.2
    # via pytest-cov
distlib==0.3.3
    # via virtualenv
distro==1.6.0
    # via ruyaml
dlint==0.12.0
    # via -r requirements-dev.in
dparse==0.5.1
    # via safety
entrypoints==0.3
    # via flakehell
eradicate==2.0.0
    # via flake8-eradicate
execnet==1.9.0
    # via pytest-xdist
factory-boy==3.2.1
    # via -r requirements-dev.in
faker==9.8.2
    # via
    #   factory-boy
    #   faker-enum
faker-enum==0.0.2
    # via -r requirements-dev.in
filelock==3.4.0
    # via virtualenv
flake8==3.9.2
    # via
    #   -r requirements-dev.in
    #   dlint
    #   flake8-annotations
    #   flake8-bugbear
    #   flake8-comprehensions
    #   flake8-debugger
    #   flake8-docstrings
    #   flake8-eradicate
    #   flake8-expression-complexity
    #   flake8-markdown
    #   flake8-mutable
    #   flake8-polyfill
    #   flake8-pytest
    #   flake8-simplify
    #   flake8-typing-imports
    #   flake8-use-fstring
    #   flakehell
    #   pep8-naming
flake8-aaa==0.12.0
    # via -r requirements-dev.in
flake8-annotations==2.7.0
    # via -r requirements-dev.in
flake8-annotations-complexity==0.0.6
    # via -r requirements-dev.in
flake8-bugbear==21.9.2
    # via -r requirements-dev.in
flake8-comprehensions==3.7.0
    # via -r requirements-dev.in
flake8-debugger==4.0.0
    # via -r requirements-dev.in
flake8-docstrings==1.6.0
    # via -r requirements-dev.in
flake8-eradicate==1.2.0
    # via -r requirements-dev.in
flake8-expression-complexity==0.0.9
    # via -r requirements-dev.in
flake8-fixme==1.1.1
    # via -r requirements-dev.in
flake8-markdown==0.3.0
    # via -r requirements-dev.in
flake8-mutable==1.2.0
    # via -r requirements-dev.in
flake8-plugin-utils==1.3.2
    # via flake8-pytest-style
flake8-polyfill==1.0.2
    # via pep8-naming
flake8-pytest==1.3
    # via -r requirements-dev.in
flake8-pytest-style==1.5.1
    # via -r requirements-dev.in
flake8-simplify==0.14.2
    # via -r requirements-dev.in
flake8-typing-imports==1.11.0
    # via -r requirements-dev.in
flake8-use-fstring==1.3
    # via -r requirements-dev.in
flake8-variables-names==0.0.4
    # via -r requirements-dev.in
flakehell==0.9.0
    # via -r requirements-dev.in
gitdb==4.0.9
    # via
    #   -c docs/requirements.txt
    #   gitpython
gitpython==3.1.24
    # via
    #   -c docs/requirements.txt
    #   bandit
identify==2.3.6
    # via pre-commit
idna==3.3
    # via
    #   -c docs/requirements.txt
    #   -c requirements.txt
    #   requests
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via
    #   -r requirements-dev.in
    #   pylint
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mypy==0.910
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
nodeenv==1.6.0
    # via pre-commit
packaging==21.2
    # via
    #   -c docs/requirements.txt
    #   dparse
    #   pytest
    #   safety
pathspec==0.9.0
    # via
    #   black
    #   yamllint
pbr==5.7.0
    # via stevedore
pep517==0.12.0
    # via pip-tools
pep8-naming==0.12.1
    # via -r requirements-dev.in
pip-tools==6.4.0
    # via -r requirements-dev.in
platformdirs==2.4.0
    # via
    #   black
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.15.0
    # via -r requirements-dev.in
py==1.11.0
    # via
    #   pytest
    #   pytest-forked
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-debugger
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.3.1
    # via
    #   autoflake
    #   flake8
pygments==2.10.0
    # via
    #   -c docs/requirements.txt
    #   flakehell
pylint==2.11.1
    # via -r requirements-dev.in
pyparsing==2.4.7
    # via
    #   -c docs/requirements.txt
    #   packaging
pyprojroot==0.2.0
    # via autoimport
pytest==6.2.5
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-forked
    #   pytest-pythonpath
    #   pytest-xdist
pytest-cov==3.0.0
    # via -r requirements-dev.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-pythonpath==0.7.3
    # via -r requirements-dev.in
pytest-xdist==2.4.0
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via
    #   -c docs/requirements.txt
    #   faker
pyyaml==6.0
    # via
    #   -c docs/requirements.txt
    #   bandit
    #   dparse
    #   pre-commit
    #   yamllint
regex==2021.11.10
    # via black
requests==2.26.0
    # via
    #   -c docs/requirements.txt
    #   -c requirements.txt
    #   safety
ruyaml==0.20.0
    # via yamlfix
safety==1.10.3
    # via -r requirements-dev.in
sh==1.14.2
    # via autoimport
six==1.16.0
    # via
    #   -c docs/requirements.txt
    #   -c requirements.txt
    #   asttokens
    #   flake8-debugger
    #   python-dateutil
    #   virtualenv
smmap==5.0.0
    # via
    #   -c docs/requirements.txt
    #   gitdb
snowballstemmer==2.1.0
    # via pydocstyle
stevedore==3.5.0
    # via bandit
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   dparse
    #   flakehell
    #   mypy
    #   pre-commit
    #   pylint
    #   pytest
tomli==1.2.2
    # via
    #   black
    #   coverage
    #   pep517
typing-extensions==4.0.0
    # via
    #   -c docs/requirements.txt
    #   -c requirements.txt
    #   astroid
    #   black
    #   gitpython
    #   mypy
    #   pylint
urllib3==1.26.7
    # via
    #   -c docs/requirements.txt
    #   -c requirements.txt
    #   flakehell
    #   requests
virtualenv==20.10.0
    # via pre-commit
wheel==0.37.0
    # via pip-tools
wrapt==1.13.3
    # via
    #   -c requirements.txt
    #   astroid
yamlfix==0.7.2
    # via -r requirements-dev.in
yamllint==1.26.3
    # via -r requirements-dev.in

# The following packages are considered to be unsafe in a requirements file:
pip==21.3.1
    # via pip-tools
setuptools==59.1.1
    # via
    #   astroid
    #   flake8-annotations-complexity
    #   flake8-expression-complexity
    #   flake8-variables-names
    #   pip-tools
    #   ruyaml
    #   safety
    #   yamllint
